+++ packages/ui/src/core/tag-field/tag-field.component.tsx
@@ -32,6 +32,8 @@ const TagField = intrinsicComponent<TagFieldProps, HTMLDivElement>(
       error,
       hint,
       loading,
+      disableOnEnter,
+      alwaysShowSuggestedTags = false,
       background = 'primary',
       getTagLabel = (tag: TagType): string => tag as string,
       getTagValue = (tag: TagType): string => tag as string,
@@ -51,7 +53,7 @@ const TagField = intrinsicComponent<TagFieldProps, HTMLDivElement>(
         (suggestion) => !existingLabels.includes(getTagLabel(suggestion).toLowerCase())
       );
 
-      return suggestionsFilter(filteredItems, userInput, getTagLabel);
+      return suggestionsFilter(filteredItems, userInput, getTagLabel, alwaysShowSuggestedTags);
     }, [userInput, suggestedTags, existingLabels, suggestionsFilter]);
 
     const handleTagAdd = (item: any, type: AddTagTypesType): void => {
@@ -65,7 +67,7 @@ const TagField = intrinsicComponent<TagFieldProps, HTMLDivElement>(
     };
 
     const handleUserInputKeyDown = (event: React.KeyboardEvent<HTMLInputElement>): void => {
-      if (event.key === 'Enter' && userInput) {
+      if (event.key === 'Enter' && userInput && !disableOnEnter) {
         event.preventDefault();
         handleTagAdd(userInput, AddTagType.UserInput);
         setUserInput('');
@@ -173,6 +175,8 @@ TagField.propTypes = {
   error: PT.bool,
   hint: PT.node,
   loading: PT.bool,
+  disableOnEnter: PT.bool,
+  alwaysShowSuggestedTags: PT.bool,
   getTagValue: PT.func,
   getTagLabel: PT.func,
   suggestionsFilter: PT.func,